refactor: extract delegate child runtime helpers from coordinator
